airflow/requirements.txt
# Astro Runtime includes the following pre-installed providers packages: https://docs.astronomer.io/astro/runtime-image-architecture#provider-packages
dbt-core~=1.8.0
astronomer-cosmos~=1.4.0
astronomer-cosmos[dbt-bigquery]~=1.4.0
apache-airflow-providers-google~=10.9.0
apache-airflow-providers-airbyte~=3.3.2